Changed Algorithms class to Movement, added functionality for B and B' movement
I make the executive decision of creating a separate class called Algorithms which will extend the Movement class. This is more fitting because what was the Algorithms class (which is now Movement) only really performed movements- not actual algorithms. Given the current size of the Movement class, it will keep the code a lot more organized by just creating a separate class that extends Movement for the algorithms.

I also added B and B' movements. I had to create two instances of the temporary values because of I categorized edge/corner pieces- not a huge deal, just slightly harder to visualize.